requirements.txt
python-dotenv>=0.15.0
numpy>=1.20.1
PyYAML>=5.4.1
opencv-python>=4.5.1.48